fix: Move main.go under dkw
