+++ .gitignore
@@ -19,4 +19,4 @@
 tmp/
 
 # Built binary
-/serve
+/dkw
