--- cmd/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "dreamkast-weaver/cmd/dkw"
-
-func main() {
-	dkw.Execute()
-}
